Remove "--enable-developer"
For one, it was a poor imitation of the automake-standard
"--enable-maintainer-mode" (A.K.A. "AM_MAINTAINER_MODE").

So why not replace it with --enable-maintainer-mode?  Search the web (and
even the automake manual) for "AM_MAINTAINER_MODE", and you will find no
shortage of arguments why it's bad.

Instead, simply use AM_MISSING_PROG to gracefully display an error message
for if the user tries to do something requiring convert or lynx without
having them installed.