Clean up intermediate file suffixes.
I was having a hell of a time trying to make sense of what a ".htm" file
was; where .in files sat, et c.

 - .in      : A file to be edited by ./configure
 - .fin     : A file to be edited by the Makefile (in a similar manner
              to .in)
 - .html    : A final HTML file
 - .tmp     : A temporary file.
 - .$X.part : A fragment of the middle of a .$X file.
 - .$X.head : A fragment of the head of a .$X file.
 - .$X.tail : A fragment of the tail of a .$X file.

I made up the ".fin"; it's like ".in", but with "f" prefixed because
they are primarily page-specific edits to "[f]ooter.html.part".  It's
not a great name, but it's better than the status quo.  I originally
had it as just another ".in" (eg. "index.html.in.in"), but that was
confusing (but had precedent; remember when the automake filename was
"Makefile.in.in" instead of "Makefile.am"?).

That is, the file extension indicates what processes it.  To facilitate
this clearer division of responsibilities, I moved the SCREENSHOT_URL
substitution into configure.ac, and moved mk_screenshots.html.bash into the
Makefile.